Skip the use of intermediate named pipes
By switching to FFmpeg's [concat demuxer][1] instead of the
[concat protocol][2], we can simpliy the script with only one `ffmpeg`
invocation, and no need to intermediate named pipes.

Note, this is only valid since all input files are (assumed to be) the
same format, which is the case when joing Cycliq files from a single
session (which is what this script is all about).

[1]: https://trac.ffmpeg.org/wiki/Concatenate#demuxer
[2]: https://trac.ffmpeg.org/wiki/Concatenate#protocol